refactor: ODM tests
