+++ tests/Batch/ImportTest.php
@@ -36,7 +36,7 @@ class ImportTest extends TestCase
 
         $this->assertIsArray($result);
         $this->assertArrayHasKey('created', $result);
-        $this->assertEquals(21, $result['created']);
+        $this->assertEquals(22, $result['created']);
     }
 
     public function testImportFromJsonDocumentsThrowServerExceptionOnNonExistentCollection()
@@ -81,7 +81,7 @@ class ImportTest extends TestCase
 
         $this->assertIsArray($result);
         $this->assertArrayHasKey('created', $result);
-        $this->assertEquals(21, $result['created']);
+        $this->assertEquals(22, $result['created']);
     }
 
     public function testImportFromArrayDocumentsThrowServerExceptionOnNonExistentCollection()
