+++ tests/Collection/Index/PrimaryIndexTest.php
@@ -1,7 +1,7 @@
 <?php
 
 
-namespace Unit\Collection\GeneralIndex;
+namespace Unit\Collection\Index;
 
 use ArangoDB\Exceptions\IndexException;
 use Unit\TestCase;
