+++ tests/Document/EdgeTest.php
@@ -13,7 +13,7 @@ use ArangoDB\Validation\Exceptions\MissingParameterException;
 
 class EdgeTest extends DocumentTestCase
 {
-    public function getAttributes($withDescriptors = false)
+    public function getAttributes($withDescriptors = false): array
     {
         $fake = parent::getAttributes($withDescriptors);
         $edgeDescriptors = [
