+++ tests/Validation/Admin/Task/TaskValidatorTest.php
@@ -11,7 +11,7 @@ use ArangoDB\Validation\Exceptions\MissingParameterException;
 
 class TaskValidatorTest extends TestCase
 {
-    public function getOptions()
+    public function getOptions(): array
     {
         return [
             'params' => [
